+++ changelog.md
@@ -4,7 +4,9 @@
 
 ### All changes:
 - Bump hahomematic to 2023.1.3
-- 
+  - Unifiy event parameters
+  - Refactor entity.py for better event support
+- Reassign event parameters in control_unit
 
 # Version 1.26.1 (2022-01-09)
 ### New features